+++ packages/components/src/Form/Fields/Field.tsx
@@ -25,7 +25,7 @@
  */
 
 import React, { FunctionComponent, ReactNode } from 'react'
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
 import { ResponsiveValue, TLengthStyledSystem } from 'styled-system'
 import omit from 'lodash/omit'
 import pick from 'lodash/pick'
@@ -124,7 +124,12 @@ const FieldLayout: FunctionComponent<FieldPropsInternal> = ({
 
   return (
     <div className={className}>
-      <Label htmlFor={id} fontWeight={labelFontWeight} fontSize={labelFontSize}>
+      <Label
+        fontSize={labelFontSize}
+        fontWeight={labelFontWeight}
+        htmlFor={id}
+        id={`${id}-labeledby`}
+      >
         {label}
         {required && <RequiredStar />}
       </Label>
@@ -147,12 +152,24 @@ FieldDetail.defaultProps = {
 const InputArea = styled.div``
 const MessageArea = styled.div``
 
+const fieldLabelCSS = (inline?: boolean) =>
+  inline
+    ? css`
+        text-align: right;
+        line-height: ${inputHeight};
+        justify-self: end;
+        height: ${inputHeight};
+        padding-right: ${({ theme }) => theme.space.small};
+      `
+    : css`
+        padding-bottom: ${({ theme }) => theme.space.xsmall};
+      `
+
 export const Field = styled(FieldLayout)<FieldPropsInternal>`
   height: fit-content;
   width: ${({ width }) => width || 'fit-content'};
   align-items: left;
   justify-content: space-between;
-  margin-bottom: ${({ theme }) => theme.space.xsmall};
 
   display: grid;
   grid-template-areas: ${({ inline }) =>
@@ -165,28 +182,16 @@ export const Field = styled(FieldLayout)<FieldPropsInternal>`
   ${InputArea} {
     display: flex;
     grid-area: input;
+    align-items: center;
   }
 
   ${MessageArea} {
     grid-area: messages;
   }
 
-  ${Label} {
+  & > ${Label} {
     grid-area: label;
-
-    ${({ inline, theme }) =>
-      inline
-        ? `
-      text-align: right;
-      line-height: ${inputHeight};
-      justify-self: end;
-      height: ${inputHeight};
-      padding-right: ${theme.space.small};
-      `
-        : `
-        padding-bottom: ${theme.space.xsmall};
-
-      `}
+    ${({ inline }) => fieldLabelCSS(inline)}
   }
 
   ${FieldDetail} {
@@ -195,10 +200,8 @@ export const Field = styled(FieldLayout)<FieldPropsInternal>`
 
     ${({ inline, theme }) =>
       inline &&
-      `
-        align-self: center;
-        padding-left: ${theme.space.small};
-      `}
+      ` align-self: center;
+        padding-left: ${theme.space.small}; `}
   }
 
   ${ValidationMessage} {