+++ packages/playground/src/index.tsx
@@ -26,13 +26,30 @@
 
 import React from 'react'
 import ReactDOM from 'react-dom'
+import { InterpolationValue } from 'styled-components'
+import { theme } from '@looker/design-tokens'
 import { ComponentsProvider } from '@looker/components'
-import { TestPopovers } from './Popovers/Testing'
+import { FieldsDemo } from './Form/FieldsDemo'
+import { FieldsetDemo } from './Form/FieldsetDemo'
+import { SpaceExamples } from './Layout/SpaceExamples'
+
+const resetTheme = {
+  ...theme,
+  reset: () => {
+    const reset: InterpolationValue = {
+      margin: 0,
+    }
+
+    return reset
+  },
+}
 
 const App: React.FC = () => {
   return (
-    <ComponentsProvider>
-      <TestPopovers />
+    <ComponentsProvider theme={resetTheme}>
+      <SpaceExamples />
+      <FieldsetDemo />
+      <FieldsDemo />
     </ComponentsProvider>
   )
 }